requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
aiosqlite==0.20.0
asitop==0.0.24
attrs==24.2.0
ble-serial==2.7.1
bleak==0.21.1
blessed==1.20.0
braceexpand==0.1.7
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
coloredlogs==15.0.1
dashing==0.1.0
discord.py==2.4.0
frozenlist==1.5.0
humanfriendly==10.0
idna==3.7
import_expression==2.2.1.post1
jishaku==2.6.0
multidict==6.1.0
numpy==1.26.4
opencv-python==4.9.0.80
propcache==0.2.0
psutil==5.9.8
pyobjc-core==9.2
pyobjc-framework-Cocoa==9.2
pyobjc-framework-CoreBluetooth==9.2
pyobjc-framework-libdispatch==9.2
python-dotenv==1.0.1
requests==2.31.0
six==1.16.0
tabulate==0.9.0
typing_extensions==4.12.2
urllib3==2.2.1
wcwidth==0.2.13
yarl==1.17.0
sqlalchemy
alembic